+ gamelog % origin children bug - unused main() variables + block definitions
The game now includes a gamelog feature (the same thing as the error feature, but the text file is a different name.)

The origin children allocation bug is now gone. The problem was that the origin block had un-initialized children pointers.
I created a function that generates an origin block so that these types of mistakes can be avoided in the future.
Origin is now a pointer to malloc'd memory now. This is for sake of consistency in the program
All blocks, including origin, will now be allocated in memory EXCLUSIVELY. There will be no hard- coded blocks in this program.
Origin is now also added to the list of blocks that block_collector keeps. It will be freed at the end of program execution.

Block definitions were adde for default elevation, origin level, and child orientation inside the parent.

error.txt now shows "Seconds = 1.234" instead of the incorrect "Ticks = 1.234".

Fixed the error.txt bug where it would always warn you that the block_collector didn't clean anything up.
This is fixed because origin is generated. So there is always at least one block allocated in memory (and it is checked into the linked block list).