every block has a pointer to a texture. Textures are only updated when necessary.
A problem that will have to be addressed evetually is that, if left unchecked, the program will generate textures for (possibly) a very very large number of blocks.
The program could allocate textures for every block the user generates (and could make the graphics run out of memory).
This is obviously a problem.