+ terrain generation + random floating point number generation functions
terrain generation is being worked on. it is by no means done yet. use q to test the terrain generator (all it currently does it change a few pixels on the edges of the camera's target block.
terrain still needs to be made to:
- check its neighbors (for continuity)
- check its parent    (for continuity)
- check its children  (for continuity)
but this can all be done once the recursive function is done.