^ block_generate_terrain_recursive_call() but there are still discontinuities.
I'm pretty sure I know why the discontinuities are there. they are on the boundaries of sub-squares. They exist because the edges of sub squares do not look at their neighbors. only the corners. therefore, the edge generation will need to be looking at the neighbors when calculating the edge points. This should solve the discontinuity problem.