all blocks are now being generated with block_generate_terrain(). continuity is still an issue.
continuity is an issue because each block's terrain is being created as the block itself is being created.
when the block is being created, all of its neighbors are null by default.
what needs to happen is for the neighbors to UPDATE after the block is created, but before the terrain is generated. IT WOULD BE NICE TO JUST DO THIS IN THE TERRAIN GENERATION FUNCTION ITSELF. so when the terrain is being generated, regardless of anything, neighbors will be updated. NOT CREATED. just updated.
I will need to write a function (or add a mode to the existing function) to simply CHECK if a neighbor exists. This check MUST NOT create additional blocks (because this would create a never-ending recursive function, and I do not wish to tangle with those demons).
too long; didn't read:
I need to make terrain generator update the neighbors of the block it is trying to generate before it generated it. It needs to do an update that does NOT create the neighbors. just updates the main block's neighbor pointers to point at its neighbors IF they exist.
wow. that was not all that much shorter.